# alfacrm_neuroland/docs/swagger_documentation.yaml
swagger: '2.0'
info:
  title: Neuroland API
  description: Документация для приложения проекта Neuroland
  contact:
    email: devappbroseph@gmail.com
  license:
    name: MIT License
  version: v1
host: 127.0.0.1:8000
schemes:
  - http
basePath: /api
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Basic:
    type: basic
security:
  - Basic: []
paths:
...
  /users/{id}/:
    get:
      operationId: users_read
      description: ''
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/AlfaCRMUserList'
      tags:
        - users
    put:
      operationId: users_update
      description: ''
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/AlfaCRMUserList'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/AlfaCRMUserList'
      tags:
        - users
    patch:
      operationId: users_partial_update
      description: ''
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/AlfaCRMUserList'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/AlfaCRMUserList'
      tags:
        - users
    delete:
      operationId: users_delete
      description: ''
      parameters: []
      responses:
        '204':
          description: ''
      tags:
        - users
    parameters:
      - name: id
        in: path
        description: A unique integer value identifying this Пользователь.
        required: true
        type: integer
...
definitions:
...
  AlfaCRMUserList:
    required:
      - name
      - phone
      - email
      - city
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      name:
        title: Имя пользователя
        description: Укажите имя
        type: string
        maxLength: 150
        minLength: 1
      phone:
        title: Номер телефона пользователя
        description: Укажите номер телефона пользователя
        type: string
        maxLength: 128
        minLength: 1
      email:
        title: Адрес электронной почты
        description: Укажите Email
        type: string
        format: email
        maxLength: 254
        minLength: 1
      avatar:
        title: Avatar
        type: string
        readOnly: true
        x-nullable: true
        format: uri
      city:
        title: City
        type: integer
      city_name:
        title: City name
        type: string
        readOnly: true
...